+++ src/choose/index.ts
@@ -1,6 +1,4 @@
 export * from "./";
 export * from "./args";
-export * from "./blocks";
 export * from "./cmd";
-export * from "./form_submission";
 export * from "./slash_command";
