+++ src/cmd/index.ts
@@ -1,3 +1,4 @@
+export * from "../choose/cmd";
 export * from "../ezpr/cmd";
 export * from "../help/cmd";
 export * from "./getNameByID";
