+++ src/help/args.ts
@@ -1,7 +1,7 @@
 import { Block, KnownBlock } from "@slack/bolt";
 
-import { error, ezprHelp, helpUsage } from "../help";
-import helpOverview from "../help/overview.json";
+import { error, ezprHelp, helpUsage } from ".";
+import helpOverview from "./overview.json";
 
 export class HelpArguments {
   topic: string;
