+++ src/help/index.ts
@@ -1,4 +1,5 @@
 export * from "./";
+export * from "./args";
 export * from "./blocks";
 export * from "./cmd";
 export * from "./slash_command";
